update packages, converted to esmodule
