+++ jquery.sparkline.js
@@ -2,7 +2,7 @@
 *
 * jquery.sparkline.js
 *
-* v@VERSION@
+* v3.2.0
 * (c) Splunk, Inc
 * Contact: Gareth Watts (gareth@splunk.com)
 * http://omnipotent.net/jquery.sparkline/
@@ -202,7 +202,7 @@
 
 /*jslint regexp: true, browser: true, jquery: true, white: true, nomen: false, plusplus: false, maxerr: 500, indent: 4 */
 
-var jQuery = require('jquery')
+import jQuery from 'jquery';
 window.jQuery = jQuery
 
 (function(document, Math, undefined) { // performance/minified-size optimization
